+++ command/start/start.go
@@ -78,9 +78,9 @@ func NewCommand(home string, docker client.CommonAPIClient, output terminal.Outp
 					continue
 				}
 
-				// make sure the app is not disabled
-				if c.Labels[containerlabels.Disabled] == "true" {
-					output.Info(fmt.Sprintf("%s is disabled, skipping", hostname))
+				// make sure the app is not suspended
+				if c.Labels[containerlabels.Suspended] == "true" {
+					output.Info(fmt.Sprintf("%s is suspended, skipping", hostname))
 					continue
 				}