+++ pkg/dockerexec/dockerexec.go
@@ -23,3 +23,23 @@ func Connect(r io.Reader, w io.Writer, user, container, shell string) error {
 
 	return c.Run()
 }
+
+// Exec is a helper package to specifically look up the path to the docker binary and execute
+// an exec command into a container.
+func Exec(r io.Reader, w io.Writer, user, container string, shell ...string) error {
+	p, err := exec.LookPath(File)
+	if err != nil {
+		return err
+	}
+
+	args := []string{"exec", "--env", "TERM=xterm-256color", "--user", user, "--interactive", "--tty", container}
+	args = append(args, shell...)
+
+	c := exec.Command(p, args...)
+
+	c.Stdin = r
+	c.Stderr = w
+	c.Stdout = w
+
+	return c.Run()
+}